// FRONTEND/chatbot-admin-panel/src/services/api/schemas/webchatUIConfig.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * Mi Chatbot IA Personalizable
 * Un backend API en Python (FastAPI) para un chatbot avanzado con IA, configurable, capaz de conectarse a diversas fuentes de contexto y con gestión de permisos para administradores.
 * OpenAPI spec version: 0.2.0
 */
import type { WebchatUIConfigBotDescription } from './webchatUIConfigBotDescription';
import type { WebchatUIConfigComposerPlaceholder } from './webchatUIConfigComposerPlaceholder';
import type { WebchatUITheme } from './webchatUITheme';
import type { WebchatUIConfigShowBetaBadge } from './webchatUIConfigShowBetaBadge';
import type { WebchatUIConfigFooterEnabled } from './webchatUIConfigFooterEnabled';
import type { WebchatUIConfigFooterText } from './webchatUIConfigFooterText';
import type { WebchatUIConfigFooterLink } from './webchatUIConfigFooterLink';
import type { WebchatUIConfigInitialState } from './webchatUIConfigInitialState';
import type { WebchatUIConfigInitialStateText } from './webchatUIConfigInitialStateText';
import type { WebchatUIConfigAvatarImageUrl } from './webchatUIConfigAvatarImageUrl';
import type { WebchatUIConfigFloatingButtonImageUrl } from './webchatUIConfigFloatingButtonImageUrl';
import type { WebchatUIConfigThemeMode } from './webchatUIConfigThemeMode';

export interface WebchatUIConfig {
  botName: string;
  botDescription?: WebchatUIConfigBotDescription;
  composerPlaceholder?: WebchatUIConfigComposerPlaceholder;
  theme: WebchatUITheme;
  showBetaBadge?: WebchatUIConfigShowBetaBadge;
  footerEnabled?: WebchatUIConfigFooterEnabled;
  footerText?: WebchatUIConfigFooterText;
  footerLink?: WebchatUIConfigFooterLink;
  initialState?: WebchatUIConfigInitialState;
  initialStateText?: WebchatUIConfigInitialStateText;
  avatarImageUrl?: WebchatUIConfigAvatarImageUrl;
  floatingButtonImageUrl?: WebchatUIConfigFloatingButtonImageUrl;
  themeMode?: WebchatUIConfigThemeMode;
}